Added validation and server side api routing
