+++ client/src/screens/AddJob.tsx
@@ -1,274 +1,10 @@
 import './AddJob.scss';
-import React, { useState } from 'react';
+import AddJobForm from '../components/AddJob/AddJobForm';
 
 function AddJob() {
-    const [formData, setFormData] = useState({
-        name: '',
-        company: '',
-        industry: '',
-        description: '',
-        skills: '',
-        responsibilities: '',
-        city: '',
-        state: '',
-        employmentType: '',
-        wage: '',
-        workType: '',
-        website: '',
-        email: '',
-        phone: '',
-      });
-    
-      const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => {
-        const { name, value } = e.target;
-        setFormData({
-          ...formData,
-          [name]: value,
-        });
-      };
-    
-      const handleSubmit = (e: React.FormEvent) => {
-        e.preventDefault();
-        console.log('Form submitted:', formData);
-        // Add logic to handle form submission (e.g., API call)
-      };
-
-  return (
-    <div className="addJob">
-      <h1>Add a New Job</h1>
-      <form onSubmit={handleSubmit}>
-        <div className="form-group">
-          <label htmlFor="name">Job Title</label>
-          <input
-            type="text"
-            id="name"
-            name="name"
-            value={formData.name}
-            onChange={handleChange}
-            placeholder="Enter job title"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="company">Company Name</label>
-          <input
-            type="text"
-            id="company"
-            name="company"
-            value={formData.company}
-            onChange={handleChange}
-            placeholder="Enter company name"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="industry">Industry</label>
-          <input
-            type="text"
-            id="industry"
-            name="industry"
-            value={formData.industry}
-            onChange={handleChange}
-            placeholder="Enter industry name"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="description">Job Description</label>
-          <textarea
-            id="description"
-            name="description"
-            value={formData.description}
-            onChange={handleChange}
-            placeholder="Enter job description"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="skills">Skills</label>
-          <input
-            type="text"
-            id="skills"
-            name="skills"
-            value={formData.skills}
-            onChange={handleChange}
-            placeholder="Enter skills list"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="responsibilities">Responsibilities</label>
-          <textarea
-            id="responsibilities"
-            name="responsibilities"
-            value={formData.responsibilities}
-            onChange={handleChange}
-            placeholder="Enter job responsibilities"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="city">City</label>
-          <input
-            type="text"
-            id="city"
-            name="city"
-            value={formData.city}
-            onChange={handleChange}
-            placeholder="Enter city where job is located"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="state">State</label>
-          <select
-            id="state"
-            name="state"
-            value={formData.state}
-            onChange={handleChange}
-            required
-          >
-          <option value="" disabled>
-              State
-            </option>
-            <option value="AL">Alabama</option>
-            <option value="AK">Alaska</option>
-            <option value="AZ">Arizona</option>
-            <option value="AR">Arkansas</option>
-            <option value="CA">California</option>
-            <option value="CO">Colorado</option>
-            <option value="CT">Connecticut</option>
-            <option value="DE">Delaware</option>
-            <option value="FL">Florida</option>
-            <option value="GA">Georgia</option>
-            <option value="HI">Hawaii</option>
-            <option value="ID">Idaho</option>
-            <option value="IL">Illinois</option>
-            <option value="IN">Indiana</option>
-            <option value="IA">Iowa</option>
-            <option value="KS">Kansas</option>
-            <option value="KY">Kentucky</option>
-            <option value="LA">Louisiana</option>
-            <option value="ME">Maine</option>
-            <option value="MD">Maryland</option>
-            <option value="MA">Massachusetts</option>
-            <option value="MI">Michigan</option>
-            <option value="MN">Minnesota</option>
-            <option value="MS">Mississippi</option>
-            <option value="MO">Missouri</option>
-            <option value="MT">Montana</option>
-            <option value="NE">Nebraska</option>
-            <option value="NV">Nevada</option>
-            <option value="NH">New Hampshire</option>
-            <option value="NJ">New Jersey</option>
-            <option value="NM">New Mexico</option>
-            <option value="NY">New York</option>
-            <option value="NC">North Carolina</option>
-            <option value="ND">North Dakota</option>
-            <option value="OH">Ohio</option>
-            <option value="OK">Oklahoma</option>
-            <option value="OR">Oregon</option>
-            <option value="PA">Pennsylvania</option>
-            <option value="RI">Rhode Island</option>
-            <option value="SC">South Carolina</option>
-            <option value="SD">South Dakota</option>
-            <option value="TN">Tennessee</option>
-            <option value="TX">Texas</option>
-            <option value="UT">Utah</option>
-            <option value="VT">Vermont</option>
-            <option value="VA">Virginia</option>
-            <option value="WA">Washington</option>
-            <option value="WV">West Virginia</option>
-            <option value="WI">Wisconsin</option>
-            <option value="WY">Wyoming</option>
-          </select>
-        </div>
-        <div className="form-group">
-          <label htmlFor="employmentType">Employment Type</label>
-          <select
-            id="employmentType"
-            name="employmentType"
-            value={formData.employmentType}
-            onChange={handleChange}
-            required
-          >
-          <option value="" disabled>
-              Employment type
-            </option>
-            <option value="Part-Time">Part-Time</option>
-            <option value="Full-Time">Full-Time</option>
-            <option value="Hybrid">Hybrid</option>
-          </select>
-        </div>
-        <div className="form-group">
-          <label htmlFor="wage">Salary</label>
-          <input
-            type="text"
-            id="wage"
-            name="wage"
-            value={formData.wage}
-            onChange={handleChange}
-            placeholder="Enter yearly salary"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="workType">Work Type</label>
-          <select
-            id="workType"
-            name="workType"
-            value={formData.workType}
-            onChange={handleChange}
-            required
-          >
-          <option value="" disabled>
-              Work type
-            </option>
-            <option value="Remote">Remote</option>
-            <option value="On-Site">On-Site</option>
-            <option value="Hybrid">Hybrid</option>
-          </select>
-        </div>
-        <div className="form-group">
-          <label htmlFor="phone">Phone Number</label>
-          <input
-            type="text"
-            id="phone"
-            name="phone"
-            value={formData.phone}
-            onChange={handleChange}
-            placeholder="Enter contact phone number"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="email">Email</label>
-          <input
-            type="text"
-            id="email"
-            name="email"
-            value={formData.email}
-            onChange={handleChange}
-            placeholder="Enter contact email"
-            required
-          />
-        </div>
-        <div className="form-group">
-          <label htmlFor="email">Website</label>
-          <input
-            type="text"
-            id="website"
-            name="website"
-            value={formData.website}
-            onChange={handleChange}
-            placeholder="Enter employer website"
-            required
-          />
-        </div>
-        <button type="submit">Submit</button>
-      </form>
-    </div>
-  );
+    return (
+        <AddJobForm />
+    );
 }
 
 export default AddJob;
