+++ server/api/controllers/jobs_controller.py
@@ -12,3 +12,18 @@ def get_all_jobs():
         error_msg = f'Failed to get jobs: {str(e)}'
         log.error(error_msg)
         return response.error(500, error_msg)
+
+def add_job():
+    try:
+        data = request.get_json()
+
+        required_fields = ['name', 'company', 'industry', 'description', 'skills', 'responsibilities', 'city', 'state', 'employmentType', 'wage', 'workType', 'website', 'email', 'phone']
+
+        for field in required_fields:
+            if field not in data:
+                return jsonify({'error': f'Missing required field: {field}'}), 400
+
+        return response.success(jobs.add_job(data))
+
+    except Exception as e:
+        return jsonify({'error': str(e)}), 500
